move yamlproxyconfig to sharding-proxy
