+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/config/RuleRegistry.java
@@ -28,7 +28,7 @@ import io.shardingsphere.core.metadata.ShardingMetaData;
 import io.shardingsphere.core.rule.MasterSlaveRule;
 import io.shardingsphere.core.rule.ProxyAuthority;
 import io.shardingsphere.core.rule.ShardingRule;
-import io.shardingsphere.core.yaml.proxy.YamlProxyConfiguration;
+import io.shardingsphere.proxy.yaml.YamlProxyConfiguration;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.proxy.metadata.ProxyShardingMetaData;
 import lombok.Getter;
